[SYCLomatic] Move MemberExprRewriters and TypeLocRewriters
