+++ clang/lib/DPCT/API/TypeNames.inc
