+++ clang/lib/DPCT/DPCT.cpp
@@ -11,8 +11,8 @@
 #include "AnalysisInfo.h"
 #include "AutoComplete.h"
 #include "Rewriters/CallExprRewriter.h"
-#include "MemberExprRewriter.h"
-#include "TypeLocRewriters.h"
+#include "Rewriters/MemberExprRewriter.h"
+#include "Rewriters/TypeLocRewriters.h"
 #include "Checkpoint.h"
 #include "Config.h"
 #include "CustomHelperFiles.h"
