+++ clang/lib/DPCT/Rewriters/TypeLocRewriterTemplateType.inc
