+++ clang/lib/DPCT/Rewriters/TypeLocRewriters.h
